Switch to better styling practices with flex boxes
Changed the footer and window styling to use flex boxes as opposed to the manual position fixed the height styling in place before. This will allow for more changeability and modularity in the future as variables in the CSS won't need to be changed to reflect how text wraps in the footer.